refactor(content): extract UmbPublishedPendingChangesManagerBase
Extract a generic base class for managing published pending changes
detection. Document overrides cleanDataBeforeComparison to null out
template before comparison. Element (on v18) will extend with no
overrides.
