+++ src/Umbraco.Web.UI.Client/src/packages/content/content/index.ts
@@ -6,6 +6,7 @@ export * from './controller/merge-content-variant-data.controller.js';
 export * from './global-components/index.js';
 export * from './manager/index.js';
 export * from './property-dataset-context/index.js';
+export * from './publishing/index.js';
 export * from './rollback/index.js';
 export * from './workspace/index.js';
