+++ src/Umbraco.Web.UI.Client/src/packages/content/content/publishing/index.ts
@@ -1,0 +1,3 @@
+export * from './published-pending-changes-manager-base.js';
+
+export type * from './types.js';
