+++ src/Umbraco.Web.UI.Client/src/packages/content/content/publishing/types.ts
@@ -1,0 +1,5 @@
+import type { UmbVariantId } from '@umbraco-cms/backoffice/variant';
+
+export interface UmbPublishedVariantWithPendingChanges {
+	variantId: UmbVariantId;
+}
